Customize table/grid layout in Tree view for better display of long file names.
Prior to this change, long strings in the Name column would either overlap text in the Message column or else wrap within the column and overlap the following Name (depending on whether the Name included a break character like ' ' or '-'). This CSS-only change tweaks the table layout to give more room to the Name/Message columns, prevent wrapping of text in the Name column, and enable ellipsis ('…') for truncated text.