Rewrite of the TestCategories.
To be able to better specify if a tests run with a certain auth
provider or membershipservice new TestCategories have been introduced.

The general idea is that you should be able to exclude tests that you
cannot run when using for example Windows Authentication because they
require for example to create a user.

To be able to accomplish this you should mark under what configuration
your test can run. If it can run under any config you do not need to add
any TestCategory.

The old categories have been replaced by IntegrationTest as all the
tests can be run together.