Fixed unit tests because the repository index does not require a logged in user
The unit tests failed because the repository page does not require the user to be logged in anymore.  I used the account details for the admin account as the test url instead.  There was a problem with selenium web driver clicking the remember me checkbox so I added a second check to make sure it was in the correct state and used some typing a space (" ") to make sure it was checked (or not).